arch: arm: rsu_s10: Retrieve the qspi_boot node with dts phandle
HSD #15013345463: Fix the symbols not found issue in Linux dtb since
CONFIG_OVERLAY_FS is not enabled by default in defconfig to create dtb
with __symbols__ node. Overlay is not created in linux dtb since it is
based on old downstream patches.
Fix the issue by retrieve qspi_boot node based on the rsu_handle
in Linux dtb.

Signed-off-by: Kah Jing Lee <kah.jing.lee@intel.com>