HSD #15015873090: arm: socfpga: use base10 for converting slot_num
RSU slot number console input was initially converted
from string to integer using hexadecimal/base16 format
when taken as input in console and when we query the slot
number in console we were printing it in decimal/base10
format. The issue came when number of slots was more
than 8, here the input function was expecting slot_num
in hexadecimal format and user input in decimal format
which was incorrect.

To fix this the string to integer conversion of slot
number in console was changed from base16 to base10.

Affected functions:
  * slot_get_info
  * slot_size
  * slot_priority
  * slot_erase
  * slot_program_buf
  * slot_program_factory_update_buf
  * slot_program_buf_raw
  * slot_verify_buf
  * slot_verify_buf_raw
  * slot_enable
  * slot_disable
  * slot_load
  * slot_rename
  * slot_delete

Signed-off-by: Mahesh Rao <mahesh.rao@intel.com>